test: simplify simulator API and complete test migration
