+++ .github/ci/_lib.sh
@@ -39,6 +39,7 @@ RMK_FEATURESETS=(
     "async_matrix,storage"
     "vial,host_lock,storage"
     "vial,_ble"
+    "vial,_ble,_no_usb,steno,passkey_entry"
     "split,async_matrix"
     "split,async_matrix,_ble"
     "split,vial,async_matrix"
