+++ rmk/Cargo.toml
@@ -166,8 +166,8 @@ log = ["dep:log", "trouble-host?/log"]
 
 ## Activate the embassy-time mock driver for deterministic, fast test timing.
 ## (Despite the historic name, this feature does NOT enable embassy-time/std —
-## it swaps in the manually-advanced mock clock that the test harness drives
-## via `tests/common/test_block_on.rs`.)
+## it swaps in the manually-advanced mock clock driven by
+## `rmk::sim::test_block_on`.)
 std = [
     "embassy-time/mock-driver",
     "embassy-time/generic-queue-128",
