+++ rmk/src/sim.rs
@@ -7,80 +7,78 @@
 //! electrical behavior.
 
 use std::boxed::Box;
-#[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+#[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
 use std::future::Future;
-#[cfg(feature = "rynk")]
-use std::marker::PhantomData;
-#[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+#[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
 use std::pin::Pin;
 use std::vec::Vec;
 
-#[cfg(any(not(feature = "_no_usb"), all(feature = "host", feature = "_ble")))]
+#[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
 use embassy_futures::select::{Either, select};
-#[cfg(not(feature = "_no_usb"))]
+#[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
+use embassy_futures::yield_now;
+#[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
 use embassy_sync::blocking_mutex::raw::CriticalSectionRawMutex;
-#[cfg(not(feature = "_no_usb"))]
+#[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
 use embassy_sync::signal::Signal;
 use embassy_time::Duration;
-#[cfg(any(not(feature = "_no_usb"), all(feature = "host", feature = "_ble")))]
+#[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
 use embassy_time::Timer;
-#[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+#[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
 use embedded_storage_async::nor_flash::NorFlash as AsyncNorFlash;
-#[cfg(not(feature = "_no_usb"))]
+#[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
 use futures::join;
-#[cfg(feature = "host")]
-use rmk_types::connection::ConnectionType;
-use rmk_types::connection::{ConnectionStatus, UsbState};
+#[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
+use rmk_types::connection::ConnectionStatus;
 #[cfg(not(feature = "_no_usb"))]
+use rmk_types::connection::UsbState;
+#[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
 use rmk_types::keycode::HidKeyCode;
 use rmk_types::morse::{Morse, MorsePattern, MorseProfile};
 #[cfg(feature = "rynk")]
-use rmk_types::protocol::rynk::{Cmd, RynkError, RynkMessage, command, endpoint::Endpoint};
-#[cfg(feature = "vial")]
-use rmk_types::protocol::vial::{SettingKey, ViaCommand, VialCommand, VialDynamic};
+use rmk_types::protocol::rynk::RynkMessage;
 
 #[cfg(not(feature = "_no_usb"))]
 use crate::channel::USB_REPORT_CHANNEL;
-#[cfg(any(feature = "host", not(feature = "_no_usb")))]
+#[cfg(any(feature = "host", not(feature = "_no_usb"), feature = "_ble"))]
 use crate::config::RmkConfig;
-#[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+#[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
 use crate::config::StorageConfig;
-use crate::config::{BehaviorConfig, CombosConfig, Hand, MorsesConfig, PositionalConfig};
-#[cfg(not(feature = "_no_usb"))]
+use crate::config::{BehaviorConfig, Hand, PositionalConfig};
+#[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
 use crate::core_traits::Runnable;
-#[cfg(not(feature = "_no_usb"))]
-use crate::event::AsyncEventPublisher;
-use crate::event::{AsyncPublishableEvent, KeyboardEvent};
-#[cfg(not(feature = "_no_usb"))]
+use crate::event::KeyboardEvent;
+#[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
+use crate::event::{AsyncEventPublisher, AsyncPublishableEvent, KeyboardEventPos};
+#[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
 use crate::hid::{KeyboardReport, Report};
-#[cfg(feature = "vial")]
-use crate::host::via::keycode_convert::to_via_keycode;
 use crate::input_device::rotary_encoder::Direction;
 use crate::keyboard::Keyboard;
 use crate::keyboard::combo::{Combo, ComboConfig};
 use crate::keymap::{KeyMap, KeymapData};
-use crate::state::{CONNECTION_STATUS, set_usb_state};
-#[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+#[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
+use crate::state::CONNECTION_STATUS;
+#[cfg(all(feature = "_no_usb", feature = "_ble"))]
+use crate::state::set_ble_state;
+#[cfg(not(feature = "_no_usb"))]
+use crate::state::set_usb_state;
+#[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
 use crate::storage::FLASH_OPERATION_FINISHED;
-#[cfg(all(feature = "storage", feature = "host"))]
-use crate::storage::FlashOperationMessage;
-#[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+#[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
 use crate::storage::Storage;
 use crate::types::action::{Action, EncoderAction, KeyAction};
 
 const DEFAULT_TIMEOUT: Duration = Duration::from_secs(5);
 
-/// Reset global test state that is backed by static channels.
-///
-/// Nextest isolates test cases by process for the mock time driver, but many RMK
-/// channels are intentionally static. A simulator scenario should call this at
-/// the start so queued events and reports cannot leak across scenarios in the
-/// same process.
-pub fn reset() {
+#[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
+fn reset() {
     KeyboardEvent::publisher_async().clear();
 
     CONNECTION_STATUS.lock(|c| c.set(ConnectionStatus::default()));
+    #[cfg(not(feature = "_no_usb"))]
     set_usb_state(UsbState::Configured);
+    #[cfg(all(feature = "_no_usb", feature = "_ble"))]
+    set_ble_state(rmk_types::ble::BleState::Connected);
 
     #[cfg(not(feature = "_no_usb"))]
     USB_REPORT_CHANNEL.clear();
@@ -89,6 +87,30 @@ pub fn reset() {
 
     #[cfg(feature = "storage")]
     crate::channel::FLASH_CHANNEL.clear();
+}
+
+#[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
+async fn receive_sim_report() -> Report {
+    #[cfg(not(feature = "_no_usb"))]
+    {
+        USB_REPORT_CHANNEL.receive().await
+    }
+    #[cfg(all(feature = "_no_usb", feature = "_ble"))]
+    {
+        crate::channel::BLE_REPORT_CHANNEL.receive().await
+    }
+}
+
+#[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
+fn try_receive_sim_report() -> Option<Report> {
+    #[cfg(not(feature = "_no_usb"))]
+    {
+        USB_REPORT_CHANNEL.try_receive().ok()
+    }
+    #[cfg(all(feature = "_no_usb", feature = "_ble"))]
+    {
+        crate::channel::BLE_REPORT_CHANNEL.try_receive().ok()
+    }
 }
 
 async fn build_static_keymap_with_encoder<
@@ -109,7 +131,7 @@ async fn build_static_keymap_with_encoder<
     Box::leak(Box::new(keymap))
 }
 
-#[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+#[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
 trait SimStorageRunner {
     fn run_until<'s>(
         &'s mut self,
@@ -117,7 +139,7 @@ trait SimStorageRunner {
     ) -> Pin<Box<dyn Future<Output = ()> + 's>>;
 }
 
-#[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+#[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
 impl<F, const ROW: usize, const COL: usize, const NUM_LAYER: usize, const NUM_ENCODER: usize> SimStorageRunner
     for Storage<F, ROW, COL, NUM_LAYER, NUM_ENCODER>
 where
@@ -136,20 +158,15 @@ where
     }
 }
 
-#[cfg(all(feature = "storage", not(feature = "_no_usb")))]
-async fn drain_flash_until_done(done: &Signal<CriticalSectionRawMutex, ()>, enabled: bool) {
-    if !enabled {
-        done.wait().await;
-        return;
-    }
-
+#[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
+async fn drain_flash_until_done(done: &Signal<CriticalSectionRawMutex, ()>) {
     match select(crate::channel::drain_flash_channel_for_test(), done.wait()).await {
         Either::First(_) => unreachable!("flash drain should never return"),
         Either::Second(()) => {}
     }
 }
 
-#[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+#[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
 async fn build_static_keymap_with_storage<
     F,
     const ROW: usize,
@@ -208,74 +225,26 @@ impl HandOverride {
 }
 
 #[derive(Clone, Copy, Debug)]
-pub struct SimMorseSetup {
-    patterns: &'static [(u16, Action)],
-    vial_morses: &'static [(Action, Action, Action, Action, MorseProfile)],
-    profile: Option<MorseProfile>,
-    flow_tap: Option<bool>,
-    prior_idle_ms: Option<u64>,
-}
-
-impl SimMorseSetup {
-    pub const fn new() -> Self {
-        Self {
-            patterns: &[],
-            vial_morses: &[],
-            profile: None,
-            flow_tap: None,
-            prior_idle_ms: None,
-        }
-    }
-
-    pub const fn patterns(mut self, patterns: &'static [(u16, Action)]) -> Self {
-        self.patterns = patterns;
-        self
-    }
-
-    pub const fn vial_morses(mut self, morses: &'static [(Action, Action, Action, Action, MorseProfile)]) -> Self {
-        self.vial_morses = morses;
-        self
-    }
-
-    pub const fn profile(mut self, profile: MorseProfile) -> Self {
-        self.profile = Some(profile);
-        self
-    }
-
-    pub const fn flow_tap(mut self, enable: bool) -> Self {
-        self.flow_tap = Some(enable);
-        self
-    }
-
-    pub const fn prior_idle_ms(mut self, prior_idle_ms: u64) -> Self {
-        self.prior_idle_ms = Some(prior_idle_ms);
-        self
-    }
-}
-
-impl Default for SimMorseSetup {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[derive(Clone, Copy, Debug)]
-pub struct SimKeyboardSetup<const ROW: usize, const COL: usize> {
+pub struct SimKeyboardSetup {
     key_overrides: &'static [KeymapOverride],
-    extra_key_overrides: &'static [KeymapOverride],
-    hands: Option<[[Hand; COL]; ROW]>,
     hand_overrides: &'static [HandOverride],
-    morse: Option<SimMorseSetup>,
+    morse_patterns: &'static [(u16, Action)],
+    vial_morses: &'static [(Action, Action, Action, Action, MorseProfile)],
+    morse_profile: Option<MorseProfile>,
+    morse_flow_tap: Option<bool>,
+    morse_prior_idle_ms: Option<u64>,
 }
 
-impl<const ROW: usize, const COL: usize> SimKeyboardSetup<ROW, COL> {
+impl SimKeyboardSetup {
     pub const fn new() -> Self {
         Self {
             key_overrides: &[],
-            extra_key_overrides: &[],
-            hands: None,
             hand_overrides: &[],
-            morse: None,
+            morse_patterns: &[],
+            vial_morses: &[],
+            morse_profile: None,
+            morse_flow_tap: None,
+            morse_prior_idle_ms: None,
         }
     }
 
@@ -284,52 +253,38 @@ impl<const ROW: usize, const COL: usize> SimKeyboardSetup<ROW, COL> {
         self
     }
 
-    pub const fn extra_keys(mut self, key_overrides: &'static [KeymapOverride]) -> Self {
-        self.extra_key_overrides = key_overrides;
-        self
-    }
-
-    pub const fn hands(mut self, hands: [[Hand; COL]; ROW]) -> Self {
-        self.hands = Some(hands);
-        self
-    }
-
     pub const fn hand_overrides(mut self, hands: &'static [HandOverride]) -> Self {
         self.hand_overrides = hands;
         self
     }
 
-    pub const fn morse(mut self, morse: SimMorseSetup) -> Self {
-        self.morse = Some(morse);
+    pub const fn morse_patterns(mut self, patterns: &'static [(u16, Action)]) -> Self {
+        self.morse_patterns = patterns;
+        self
+    }
+
+    pub const fn vial_morses(mut self, morses: &'static [(Action, Action, Action, Action, MorseProfile)]) -> Self {
+        self.vial_morses = morses;
         self
     }
 
     pub const fn morse_profile(mut self, profile: MorseProfile) -> Self {
-        self.morse = Some(match self.morse {
-            Some(morse) => morse.profile(profile),
-            None => SimMorseSetup::new().profile(profile),
-        });
+        self.morse_profile = Some(profile);
         self
     }
 
     pub const fn morse_flow_tap(mut self, enable: bool) -> Self {
-        self.morse = Some(match self.morse {
-            Some(morse) => morse.flow_tap(enable),
-            None => SimMorseSetup::new().flow_tap(enable),
-        });
+        self.morse_flow_tap = Some(enable);
         self
     }
 
     pub const fn morse_prior_idle_ms(mut self, prior_idle_ms: u64) -> Self {
-        self.morse = Some(match self.morse {
-            Some(morse) => morse.prior_idle_ms(prior_idle_ms),
-            None => SimMorseSetup::new().prior_idle_ms(prior_idle_ms),
-        });
+        self.morse_prior_idle_ms = Some(prior_idle_ms);
         self
     }
 }
 
-impl<const ROW: usize, const COL: usize> Default for SimKeyboardSetup<ROW, COL> {
+impl Default for SimKeyboardSetup {
     fn default() -> Self {
         Self::new()
     }
@@ -337,19 +292,15 @@ impl<const ROW: usize, const COL: usize> Default for SimKeyboardSetup<ROW, COL> 
 
 pub struct NoSimStorage;
 
-#[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+#[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
 pub struct SimStorage<F> {
     flash: F,
-    config: StorageConfig,
 }
 
-#[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+#[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
 impl<F> SimStorage<F> {
     fn new(flash: F) -> Self {
-        Self {
-            flash,
-            config: StorageConfig::default(),
-        }
+        Self { flash }
     }
 }
 
@@ -386,63 +337,33 @@ impl<const ROW: usize, const COL: usize, const NUM_LAYER: usize> SimKeyboardBuil
 impl<const ROW: usize, const COL: usize, const NUM_LAYER: usize, const NUM_ENCODER: usize, S>
     SimKeyboardBuilder<ROW, COL, NUM_LAYER, NUM_ENCODER, S>
 {
-    pub fn behavior(mut self, behavior_config: BehaviorConfig) -> Self {
-        self.behavior_config = behavior_config;
-        self
-    }
-
-    pub fn positional(mut self, positional_config: PositionalConfig<ROW, COL>) -> Self {
-        self.positional_config = positional_config;
-        self
-    }
-
-    pub fn hands(mut self, hand: [[Hand; COL]; ROW]) -> Self {
-        self.positional_config = PositionalConfig::new(hand);
-        self
-    }
-
-    pub fn hand(mut self, row: usize, col: usize, hand: Hand) -> Self {
-        self.positional_config.hand[row][col] = hand;
-        self
-    }
-
     pub fn key(mut self, layer: usize, row: usize, col: usize, action: KeyAction) -> Self {
         self.keymap[layer][row][col] = action;
         self
     }
 
-    pub fn setup(mut self, setup: SimKeyboardSetup<ROW, COL>) -> Self {
-        for key in setup.key_overrides.iter().chain(setup.extra_key_overrides) {
+    pub fn setup(mut self, setup: SimKeyboardSetup) -> Self {
+        for key in setup.key_overrides {
             self.keymap[key.layer][key.row][key.col] = key.action;
-        }
-        if let Some(hands) = setup.hands {
-            self = self.hands(hands);
         }
         for hand in setup.hand_overrides {
             self.positional_config.hand[hand.row][hand.col] = hand.hand;
         }
-        if let Some(morse) = setup.morse {
-            if let Some(enable) = morse.flow_tap {
-                self = self.morse_flow_tap(enable);
-            }
-            if let Some(prior_idle_ms) = morse.prior_idle_ms {
-                self = self.morse_prior_idle_ms(prior_idle_ms);
-            }
-            if let Some(profile) = morse.profile {
-                self = self.morse_default_profile(profile);
-            }
-            if !morse.patterns.is_empty() {
-                self = self.morse_patterns_slice(morse.patterns);
-            }
-            if !morse.vial_morses.is_empty() {
-                self = self.morses_from_vial_slice(morse.vial_morses);
-            }
+        if let Some(enable) = setup.morse_flow_tap {
+            self = self.morse_flow_tap(enable);
         }
-        self
-    }
-
-    pub fn morse_config(mut self, morse_config: MorsesConfig) -> Self {
-        self.behavior_config.morse = morse_config;
+        if let Some(prior_idle_ms) = setup.morse_prior_idle_ms {
+            self = self.morse_prior_idle_ms(prior_idle_ms);
+        }
+        if let Some(profile) = setup.morse_profile {
+            self = self.morse_default_profile(profile);
+        }
+        if !setup.morse_patterns.is_empty() {
+            self = self.morse_patterns_slice(setup.morse_patterns);
+        }
+        if !setup.vial_morses.is_empty() {
+            self = self.morses_from_vial_slice(setup.vial_morses);
+        }
         self
     }
 
@@ -456,13 +377,9 @@ impl<const ROW: usize, const COL: usize, const NUM_LAYER: usize, const NUM_ENCOD
         self
     }
 
-    pub fn morse_prior_idle_time(mut self, prior_idle_time: Duration) -> Self {
-        self.behavior_config.morse.prior_idle_time = prior_idle_time;
+    pub fn morse_prior_idle_ms(mut self, prior_idle_ms: u64) -> Self {
+        self.behavior_config.morse.prior_idle_time = Duration::from_millis(prior_idle_ms);
         self
-    }
-
-    pub fn morse_prior_idle_ms(self, prior_idle_ms: u64) -> Self {
-        self.morse_prior_idle_time(Duration::from_millis(prior_idle_ms))
     }
 
     pub fn morse(mut self, morse: Morse) -> Self {
@@ -485,32 +402,11 @@ impl<const ROW: usize, const COL: usize, const NUM_LAYER: usize, const NUM_ENCOD
         self.morse(Morse::new_from_vial(tap, hold, hold_after_tap, double_tap, profile))
     }
 
-    pub fn morses_from_vial<const NUM_MORSE: usize>(
-        mut self,
-        morses: [(Action, Action, Action, Action, MorseProfile); NUM_MORSE],
-    ) -> Self {
-        for (tap, hold, hold_after_tap, double_tap, profile) in morses {
-            self = self.morse_from_vial(tap, hold, hold_after_tap, double_tap, profile);
-        }
-        self
-    }
-
     fn morses_from_vial_slice(mut self, morses: &[(Action, Action, Action, Action, MorseProfile)]) -> Self {
         for &(tap, hold, hold_after_tap, double_tap, profile) in morses {
             self = self.morse_from_vial(tap, hold, hold_after_tap, double_tap, profile);
         }
         self
-    }
-
-    pub fn morse_patterns<const NUM_PATTERN: usize>(self, patterns: [(u16, Action); NUM_PATTERN]) -> Self {
-        self.morse(Morse {
-            actions: heapless::LinearMap::from_iter(
-                patterns
-                    .into_iter()
-                    .map(|(pattern, action)| (MorsePattern::from_u16(pattern), action)),
-            ),
-            ..Default::default()
-        })
     }
 
     fn morse_patterns_slice(self, patterns: &[(u16, Action)]) -> Self {
@@ -525,15 +421,7 @@ impl<const ROW: usize, const COL: usize, const NUM_LAYER: usize, const NUM_ENCOD
         })
     }
 
-    pub fn morse_patterns_with_profile<const NUM_PATTERN: usize>(
-        self,
-        profile: MorseProfile,
-        patterns: [(u16, Action); NUM_PATTERN],
-    ) -> Self {
-        self.morse_default_profile(profile).morse_patterns(patterns)
-    }
-
-    pub fn combo<const NUM_ACTION: usize>(
+    fn combo<const NUM_ACTION: usize>(
         mut self,
         actions: [KeyAction; NUM_ACTION],
         output: KeyAction,
@@ -585,32 +473,14 @@ impl<const ROW: usize, const COL: usize, const NUM_LAYER: usize, const NUM_ENCOD
         self
     }
 
-    pub fn combo_timeout(mut self, timeout: Duration) -> Self {
-        self.behavior_config.combo.timeout = timeout;
+    pub fn combo_timeout_ms(mut self, timeout_ms: u64) -> Self {
+        self.behavior_config.combo.timeout = Duration::from_millis(timeout_ms);
         self
     }
 
-    pub fn combo_timeout_ms(self, timeout_ms: u64) -> Self {
-        self.combo_timeout(Duration::from_millis(timeout_ms))
-    }
-
-    pub fn combo_prior_idle_time(mut self, prior_idle_time: Option<Duration>) -> Self {
-        self.behavior_config.combo.prior_idle_time = prior_idle_time;
+    pub fn one_shot_timeout_ms(mut self, timeout_ms: u64) -> Self {
+        self.behavior_config.one_shot.timeout = Duration::from_millis(timeout_ms);
         self
-    }
-
-    pub fn combo_config(mut self, combo_config: CombosConfig) -> Self {
-        self.behavior_config.combo = combo_config;
-        self
-    }
-
-    pub fn one_shot_timeout(mut self, timeout: Duration) -> Self {
-        self.behavior_config.one_shot.timeout = timeout;
-        self
-    }
-
-    pub fn one_shot_timeout_ms(self, timeout_ms: u64) -> Self {
-        self.one_shot_timeout(Duration::from_millis(timeout_ms))
     }
 
     pub fn one_shot_activate_on_keypress(mut self, activate_on_keypress: bool) -> Self {
@@ -649,14 +519,7 @@ impl<const ROW: usize, const COL: usize, const NUM_LAYER: usize, const NUM_ENCOD
         self
     }
 
-    #[cfg(all(feature = "storage", not(feature = "_no_usb")))]
-    pub fn storage(
-        self,
-    ) -> SimKeyboardBuilder<ROW, COL, NUM_LAYER, NUM_ENCODER, SimStorage<flash::InMemoryFlash<8192>>> {
-        self.storage_flash(flash::InMemoryFlash::new())
-    }
-
-    #[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+    #[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
     pub fn storage_flash<F>(self, flash: F) -> SimKeyboardBuilder<ROW, COL, NUM_LAYER, NUM_ENCODER, SimStorage<F>>
     where
         F: AsyncNorFlash + 'static,
@@ -695,23 +558,18 @@ impl<const ROW: usize, const COL: usize, const NUM_LAYER: usize, const NUM_ENCOD
     }
 }
 
-#[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+#[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
 impl<F, const ROW: usize, const COL: usize, const NUM_LAYER: usize, const NUM_ENCODER: usize>
     SimKeyboardBuilder<ROW, COL, NUM_LAYER, NUM_ENCODER, SimStorage<F>>
 where
     F: AsyncNorFlash + 'static,
 {
-    pub fn storage_config(mut self, config: StorageConfig) -> Self {
-        self.storage.config = config;
-        self
-    }
-
     pub async fn build(self) -> SimKeyboard<'static> {
         let (keymap, storage) = build_static_keymap_with_storage(
             self.keymap,
             self.encoder_map,
             self.storage.flash,
-            self.storage.config,
+            StorageConfig::default(),
             self.behavior_config,
             self.positional_config,
         )
@@ -731,32 +589,26 @@ where
 enum SimStep {
     Event(KeyboardEvent),
     Delay(Duration),
-    #[cfg(not(feature = "_no_usb"))]
+    #[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
     ExpectKeyboardState {
         modifier: u8,
         keycodes: Vec<u8>,
     },
-    #[cfg(not(feature = "_no_usb"))]
+    #[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
     ExpectReport(Report),
-    #[cfg(not(feature = "_no_usb"))]
+    #[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
     ExpectNoReport(Duration),
     #[cfg(feature = "vial")]
     VialPacket {
-        transport: ConnectionType,
         data: [u8; 32],
-        timeout: Duration,
-        expected: VialReplyExpectation,
+        expected: [u8; 32],
     },
     #[cfg(feature = "rynk")]
     RynkPacket {
-        transport: ConnectionType,
-        timeout: Duration,
         request: Vec<u8>,
         expected: Vec<u8>,
     },
-    #[cfg(all(feature = "storage", feature = "host"))]
-    ExpectFlashOperation(FlashExpectation),
-    #[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+    #[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
     WaitStorage,
     #[cfg(feature = "passkey_entry")]
     BeginPasskeyEntry,
@@ -766,42 +618,12 @@ enum SimStep {
     EndPasskeyEntry,
 }
 
-#[cfg(feature = "vial")]
-#[derive(Debug)]
-enum VialReplyExpectation {
-    Exact([u8; 32]),
-    Command(u8),
-}
-
-#[cfg(all(feature = "storage", feature = "host"))]
-#[derive(Debug)]
-enum FlashExpectation {
-    KeymapKey {
-        layer: u8,
-        row: u8,
-        col: u8,
-        action: KeyAction,
-    },
-    Encoder {
-        layer: u8,
-        idx: u8,
-        action: EncoderAction,
-    },
-    ComboTimeout(u16),
-    OneShotTimeout(u16),
-    TapInterval(u16),
-    TapCapslockInterval(u16),
-    PriorIdleTime(u16),
-    MorseDefaultProfile(MorseProfile),
-}
-
 /// Simulator for a complete keyboard device, excluding physical input and
 /// transport I/O.
 pub struct SimKeyboard<'a> {
     keyboard: Keyboard<'a>,
-    timeout: Duration,
     steps: Vec<SimStep>,
-    #[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+    #[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
     storage: Option<Box<dyn SimStorageRunner>>,
     #[cfg(feature = "host")]
     host_config: Option<RmkConfig<'static>>,
@@ -811,23 +633,17 @@ impl<'a> SimKeyboard<'a> {
     fn new(keyboard: Keyboard<'a>) -> Self {
         Self {
             keyboard,
-            timeout: DEFAULT_TIMEOUT,
             steps: Vec::new(),
-            #[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+            #[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
             storage: None,
             #[cfg(feature = "host")]
             host_config: None,
         }
     }
 
-    #[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+    #[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
     fn with_storage(mut self, storage: Box<dyn SimStorageRunner>) -> Self {
         self.storage = Some(storage);
-        self
-    }
-
-    pub fn with_timeout(mut self, timeout: Duration) -> Self {
-        self.timeout = timeout;
         self
     }
 
@@ -848,11 +664,7 @@ impl<'a> SimKeyboard<'a> {
     }
 
     pub fn delay(&mut self, ms: u64) -> &mut Self {
-        self.delay_duration(Duration::from_millis(ms))
-    }
-
-    pub fn delay_duration(&mut self, duration: Duration) -> &mut Self {
-        self.steps.push(SimStep::Delay(duration));
+        self.steps.push(SimStep::Delay(Duration::from_millis(ms)));
         self
     }
 
@@ -864,7 +676,7 @@ impl<'a> SimKeyboard<'a> {
         self.rotary(id, Direction::CounterClockwise)
     }
 
-    pub fn rotary(&mut self, id: u8, direction: Direction) -> &mut Self {
+    fn rotary(&mut self, id: u8, direction: Direction) -> &mut Self {
         self.event(KeyboardEvent::rotary_encoder(id, direction, true))
             .event(KeyboardEvent::rotary_encoder(id, direction, false))
     }
@@ -892,18 +704,18 @@ impl<'a> SimKeyboard<'a> {
         self
     }
 
-    #[cfg(not(feature = "_no_usb"))]
+    #[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
     pub fn expect_keys<const N: usize>(&mut self, keycodes: [HidKeyCode; N]) -> &mut Self {
         self.expect_keys_with_mods(0, keycodes)
     }
 
-    #[cfg(not(feature = "_no_usb"))]
+    #[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
     pub fn expect_keys_with_mods<const N: usize>(&mut self, modifier: u8, keycodes: [HidKeyCode; N]) -> &mut Self {
         let keycodes = keycodes.iter().map(|keycode| *keycode as u8).collect();
         self.expect_keyboard_state(modifier, keycodes)
     }
 
-    #[cfg(not(feature = "_no_usb"))]
+    #[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
     fn expect_keyboard_state(&mut self, modifier: u8, keycodes: Vec<u8>) -> &mut Self {
         assert!(
             keycodes.len() <= KeyboardReport::default().keycodes.len(),
@@ -914,28 +726,28 @@ impl<'a> SimKeyboard<'a> {
         self
     }
 
-    #[cfg(not(feature = "_no_usb"))]
+    #[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
     pub fn expect_only_mods(&mut self, modifier: u8) -> &mut Self {
         self.expect_keyboard_state(modifier, Vec::new())
     }
 
-    #[cfg(not(feature = "_no_usb"))]
+    #[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
     pub fn expect_all_up(&mut self) -> &mut Self {
         self.expect_keyboard_state(0, Vec::new())
     }
 
-    #[cfg(not(feature = "_no_usb"))]
+    #[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
     pub fn expect_keyboard_report(&mut self, report: KeyboardReport) -> &mut Self {
         self.expect_report(Report::KeyboardReport(report))
     }
 
-    #[cfg(not(feature = "_no_usb"))]
+    #[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
     pub fn expect_report(&mut self, report: Report) -> &mut Self {
         self.steps.push(SimStep::ExpectReport(report));
         self
     }
 
-    #[cfg(not(feature = "_no_usb"))]
+    #[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
     pub fn expect_no_report(&mut self, ms: u64) -> &mut Self {
         self.steps.push(SimStep::ExpectNoReport(Duration::from_millis(ms)));
         self
@@ -949,129 +761,31 @@ impl<'a> SimKeyboard<'a> {
     }
 
     #[cfg(feature = "vial")]
-    fn vial_packet(
-        &mut self,
-        transport: ConnectionType,
-        data: [u8; 32],
-        timeout: Duration,
-        expected: VialReplyExpectation,
-    ) -> &mut Self {
-        self.steps.push(SimStep::VialPacket {
-            transport,
-            data,
-            timeout,
-            expected,
-        });
+    fn vial_packet(&mut self, data: [u8; 32], expected: [u8; 32]) -> &mut Self {
+        self.steps.push(SimStep::VialPacket { data, expected });
         self
     }
 
-    #[cfg(all(feature = "storage", feature = "host"))]
-    pub fn expect_flash_key(&mut self, layer: u8, row: u8, col: u8, action: KeyAction) -> &mut Self {
-        self.steps
-            .push(SimStep::ExpectFlashOperation(FlashExpectation::KeymapKey {
-                layer,
-                row,
-                col,
-                action,
-            }));
-        self
-    }
-
-    #[cfg(all(feature = "storage", feature = "host"))]
-    pub fn expect_flash_encoder(&mut self, layer: u8, idx: u8, action: EncoderAction) -> &mut Self {
-        self.steps
-            .push(SimStep::ExpectFlashOperation(FlashExpectation::Encoder {
-                layer,
-                idx,
-                action,
-            }));
-        self
-    }
-
-    #[cfg(all(feature = "storage", feature = "host"))]
-    pub fn expect_flash_combo_timeout(&mut self, value: u16) -> &mut Self {
-        self.steps
-            .push(SimStep::ExpectFlashOperation(FlashExpectation::ComboTimeout(value)));
-        self
-    }
-
-    #[cfg(all(feature = "storage", feature = "host"))]
-    pub fn expect_flash_one_shot_timeout(&mut self, value: u16) -> &mut Self {
-        self.steps
-            .push(SimStep::ExpectFlashOperation(FlashExpectation::OneShotTimeout(value)));
-        self
-    }
-
-    #[cfg(all(feature = "storage", feature = "host"))]
-    pub fn expect_flash_tap_interval(&mut self, value: u16) -> &mut Self {
-        self.steps
-            .push(SimStep::ExpectFlashOperation(FlashExpectation::TapInterval(value)));
-        self
-    }
-
-    #[cfg(all(feature = "storage", feature = "host"))]
-    pub fn expect_flash_tap_capslock_interval(&mut self, value: u16) -> &mut Self {
-        self.steps
-            .push(SimStep::ExpectFlashOperation(FlashExpectation::TapCapslockInterval(
-                value,
-            )));
-        self
-    }
-
-    #[cfg(all(feature = "storage", feature = "host"))]
-    pub fn expect_flash_prior_idle_time(&mut self, value: u16) -> &mut Self {
-        self.steps
-            .push(SimStep::ExpectFlashOperation(FlashExpectation::PriorIdleTime(value)));
-        self
-    }
-
-    #[cfg(all(feature = "storage", feature = "host"))]
-    pub fn expect_flash_morse_default_profile(&mut self, profile: MorseProfile) -> &mut Self {
-        self.steps
-            .push(SimStep::ExpectFlashOperation(FlashExpectation::MorseDefaultProfile(
-                profile,
-            )));
-        self
-    }
-
-    #[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+    #[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
     pub fn wait_storage(&mut self) -> &mut Self {
         self.steps.push(SimStep::WaitStorage);
         self
     }
 
     #[cfg(feature = "rynk")]
-    fn rynk_packet(
-        &mut self,
-        transport: ConnectionType,
-        timeout: Duration,
-        request: Vec<u8>,
-        expected: Vec<u8>,
-    ) -> &mut Self {
-        self.steps.push(SimStep::RynkPacket {
-            transport,
-            timeout,
-            request,
-            expected,
-        });
+    fn rynk_packet(&mut self, request: Vec<u8>, expected: Vec<u8>) -> &mut Self {
+        self.steps.push(SimStep::RynkPacket { request, expected });
         self
     }
 
-    #[cfg(not(feature = "_no_usb"))]
+    #[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
     pub async fn run(&mut self) {
         let keyboard_done = Signal::<CriticalSectionRawMutex, ()>::new();
-        #[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+        #[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
         let storage_done = Signal::<CriticalSectionRawMutex, ()>::new();
-        #[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+        #[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
         let flash_drain_done = Signal::<CriticalSectionRawMutex, ()>::new();
         let steps = core::mem::take(&mut self.steps);
-        let timeout = self.timeout;
-        #[cfg(all(feature = "storage", feature = "host"))]
-        let drain_flash = !steps
-            .iter()
-            .any(|step| matches!(step, SimStep::ExpectFlashOperation(_)));
-        #[cfg(all(feature = "storage", not(feature = "host")))]
-        let drain_flash = true;
         #[cfg(feature = "host")]
         let protocol_config = self.host_config.as_ref();
         #[cfg(not(feature = "host"))]
@@ -1080,14 +794,14 @@ impl<'a> SimKeyboard<'a> {
         reset();
         let keymap = self.keyboard.keymap;
 
-        #[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+        #[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
         if let Some(storage) = self.storage.as_deref_mut() {
             join!(
                 Self::run_keyboard_until_done(&mut self.keyboard, &keyboard_done),
                 storage.run_until(&storage_done),
-                drain_flash_until_done(&flash_drain_done, false),
+                flash_drain_done.wait(),
                 async {
-                    Self::run_steps(keymap, steps, timeout, protocol_config).await;
+                    Self::run_steps(keymap, steps, DEFAULT_TIMEOUT, protocol_config).await;
                     keyboard_done.signal(());
                     storage_done.signal(());
                     flash_drain_done.signal(());
@@ -1102,17 +816,17 @@ impl<'a> SimKeyboard<'a> {
         join!(
             Self::run_keyboard_until_done(&mut self.keyboard, &keyboard_done),
             async {
-                Self::run_steps(keymap, steps, timeout, protocol_config).await;
+                Self::run_steps(keymap, steps, DEFAULT_TIMEOUT, protocol_config).await;
                 keyboard_done.signal(());
                 flash_drain_done.signal(());
             },
-            drain_flash_until_done(&flash_drain_done, drain_flash)
+            drain_flash_until_done(&flash_drain_done)
         );
         #[cfg(not(feature = "storage"))]
         join!(
             Self::run_keyboard_until_done(&mut self.keyboard, &keyboard_done),
             async {
-                Self::run_steps(keymap, steps, timeout, protocol_config).await;
+                Self::run_steps(keymap, steps, DEFAULT_TIMEOUT, protocol_config).await;
                 keyboard_done.signal(());
             }
         );
@@ -1127,9 +841,15 @@ impl<'a> SimKeyboard<'a> {
                 self.keyboard.held_buffer
             );
         }
+        if !self.keyboard.unprocessed_events.is_empty() {
+            panic!(
+                "simulator ended with unprocessed keyboard events: {:?}",
+                self.keyboard.unprocessed_events
+            );
+        }
     }
 
-    #[cfg(not(feature = "_no_usb"))]
+    #[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
     async fn run_keyboard_until_done(keyboard: &mut Keyboard<'_>, done: &Signal<CriticalSectionRawMutex, ()>) {
         match select(keyboard.run(), done.wait()).await {
             Either::First(_) => unreachable!("keyboard task should never return"),
@@ -1137,7 +857,7 @@ impl<'a> SimKeyboard<'a> {
         }
     }
 
-    #[cfg(not(feature = "_no_usb"))]
+    #[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
     async fn run_steps<'k>(
         keymap: &'k KeyMap<'k>,
         steps: Vec<SimStep>,
@@ -1150,53 +870,58 @@ impl<'a> SimKeyboard<'a> {
         let sender = KeyboardEvent::publisher_async();
         #[cfg(feature = "host")]
         let host_service = protocol_config.map(|config| crate::host::HostService::new(keymap, config));
+        let mut pressed_inputs = Vec::<KeyboardEventPos>::new();
 
         for (idx, step) in steps.into_iter().enumerate() {
             match step {
-                SimStep::Event(event) => match select(Timer::after(timeout), sender.publish_async(event)).await {
-                    Either::First(_) => panic!("simulator timed out publishing keyboard event at step #{idx}"),
-                    Either::Second(()) => {}
-                },
+                SimStep::Event(event) => {
+                    if event.pressed {
+                        assert!(
+                            !pressed_inputs.contains(&event.pos),
+                            "input {:?} was pressed twice without a release at step #{idx}",
+                            event.pos
+                        );
+                        pressed_inputs.push(event.pos);
+                    } else {
+                        let Some(pos) = pressed_inputs.iter().position(|pressed| *pressed == event.pos) else {
+                            panic!(
+                                "input {:?} was released without a matching press at step #{idx}",
+                                event.pos
+                            );
+                        };
+                        pressed_inputs.swap_remove(pos);
+                    }
+
+                    match select(Timer::after(timeout), sender.publish_async(event)).await {
+                        Either::First(_) => panic!("simulator timed out publishing keyboard event at step #{idx}"),
+                        Either::Second(()) => {}
+                    }
+                }
                 SimStep::Delay(duration) => {
                     Timer::after(duration).await;
                 }
                 SimStep::ExpectKeyboardState { modifier, keycodes } => {
-                    let actual = match select(Timer::after(timeout), USB_REPORT_CHANNEL.receive()).await {
+                    let actual = match select(Timer::after(timeout), receive_sim_report()).await {
                         Either::First(_) => panic!("simulator timed out waiting for keyboard report at step #{idx}"),
                         Either::Second(report) => report,
                     };
                     Self::assert_keyboard_state_eq(modifier, keycodes, actual, idx);
                 }
                 SimStep::ExpectReport(expected) => {
-                    let actual = match select(Timer::after(timeout), USB_REPORT_CHANNEL.receive()).await {
+                    let actual = match select(Timer::after(timeout), receive_sim_report()).await {
                         Either::First(_) => panic!("simulator timed out waiting for HID report at step #{idx}"),
                         Either::Second(report) => report,
                     };
                     Self::assert_report_eq(expected, actual, idx);
                 }
-                SimStep::ExpectNoReport(duration) => {
-                    match select(Timer::after(duration), USB_REPORT_CHANNEL.receive()).await {
-                        Either::First(_) => {}
-                        Either::Second(report) => {
-                            panic!("unexpected HID report at step #{idx}: {:?}", report);
-                        }
+                SimStep::ExpectNoReport(duration) => match select(Timer::after(duration), receive_sim_report()).await {
+                    Either::First(_) => {}
+                    Either::Second(report) => {
+                        panic!("unexpected HID report at step #{idx}: {:?}", report);
                     }
-                }
+                },
                 #[cfg(feature = "vial")]
-                SimStep::VialPacket {
-                    transport,
-                    data,
-                    timeout,
-                    expected,
-                } => {
-                    #[cfg(not(feature = "_ble"))]
-                    assert_ne!(
-                        transport,
-                        ConnectionType::Ble,
-                        "BLE host transactions require the `_ble` feature"
-                    );
-                    #[cfg(feature = "_ble")]
-                    let _ = transport;
+                SimStep::VialPacket { data, expected } => {
                     #[cfg(feature = "storage")]
                     FLASH_OPERATION_FINISHED.reset();
                     let service = host_service
@@ -1206,30 +931,10 @@ impl<'a> SimKeyboard<'a> {
                         Either::First(_) => panic!("simulator timed out dispatching Vial packet at step #{idx}"),
                         Either::Second(reply) => reply,
                     };
-                    match expected {
-                        VialReplyExpectation::Exact(expected) => {
-                            assert_eq!(expected, actual, "on Vial reply at step #{idx}");
-                        }
-                        VialReplyExpectation::Command(command) => {
-                            assert_eq!(command, actual[0], "on Vial reply command at step #{idx}");
-                        }
-                    }
+                    assert_eq!(expected, actual, "on Vial reply at step #{idx}");
                 }
                 #[cfg(feature = "rynk")]
-                SimStep::RynkPacket {
-                    transport,
-                    timeout,
-                    mut request,
-                    expected,
-                } => {
-                    #[cfg(not(feature = "_ble"))]
-                    assert_ne!(
-                        transport,
-                        ConnectionType::Ble,
-                        "BLE host transactions require the `_ble` feature"
-                    );
-                    #[cfg(feature = "_ble")]
-                    let _ = transport;
+                SimStep::RynkPacket { mut request, expected } => {
                     #[cfg(feature = "storage")]
                     FLASH_OPERATION_FINISHED.reset();
                     let service = host_service
@@ -1250,15 +955,7 @@ impl<'a> SimKeyboard<'a> {
                         &request[..frame_len]
                     );
                 }
-                #[cfg(all(feature = "storage", feature = "host"))]
-                SimStep::ExpectFlashOperation(expected) => {
-                    let actual = match select(Timer::after(timeout), crate::channel::FLASH_CHANNEL.receive()).await {
-                        Either::First(_) => panic!("simulator timed out waiting for flash operation at step #{idx}"),
-                        Either::Second(operation) => operation,
-                    };
-                    Self::assert_flash_operation_eq(expected, actual, idx);
-                }
-                #[cfg(all(feature = "storage", not(feature = "_no_usb")))]
+                #[cfg(all(feature = "storage", any(not(feature = "_no_usb"), feature = "_ble")))]
                 SimStep::WaitStorage => match select(Timer::after(timeout), FLASH_OPERATION_FINISHED.wait()).await {
                     Either::First(_) => panic!("simulator timed out waiting for storage write at step #{idx}"),
                     Either::Second(true) => {}
@@ -1286,83 +983,34 @@ impl<'a> SimKeyboard<'a> {
             }
         }
 
-        // Give the keyboard task a chance to consume a final event even when
-        // the timeline has no following expectation, then reject output that
-        // the scenario did not explicitly assert.
+        match select(Timer::after(timeout), async {
+            while !sender.is_empty() {
+                yield_now().await;
+            }
+        })
+        .await
+        {
+            Either::First(_) => panic!("simulator timed out draining keyboard events after the final step"),
+            Either::Second(()) => {}
+        }
+
+        // The queue becomes empty when the keyboard receives the final event;
+        // allow that in-flight processing to finish before checking state.
         Timer::after(Duration::from_millis(1)).await;
-        if let Ok(report) = USB_REPORT_CHANNEL.try_receive() {
+        if let Some(report) = try_receive_sim_report() {
             panic!(
                 "unexpected trailing HID report after final simulator step: {:?}",
                 report
             );
         }
+        assert!(
+            pressed_inputs.is_empty(),
+            "simulator ended with pressed inputs: {:?}",
+            pressed_inputs
+        );
     }
 
-    #[cfg(all(feature = "storage", feature = "host"))]
-    fn assert_flash_operation_eq(expected: FlashExpectation, actual: FlashOperationMessage, step_idx: usize) {
-        match (expected, actual) {
-            (
-                FlashExpectation::KeymapKey {
-                    layer,
-                    row,
-                    col,
-                    action,
-                },
-                FlashOperationMessage::KeymapKey {
-                    layer: actual_layer,
-                    row: actual_row,
-                    col: actual_col,
-                    action: actual_action,
-                },
-            ) => {
-                assert_eq!(
-                    (layer, row, col, action),
-                    (actual_layer, actual_row, actual_col, actual_action),
-                    "on flash operation at step #{step_idx}"
-                );
-            }
-            (
-                FlashExpectation::Encoder { layer, idx, action },
-                FlashOperationMessage::Encoder {
-                    layer: actual_layer,
-                    idx: actual_idx,
-                    action: actual_action,
-                },
-            ) => {
-                assert_eq!(
-                    (layer, idx, action),
-                    (actual_layer, actual_idx, actual_action),
-                    "on flash operation at step #{step_idx}"
-                );
-            }
-            (FlashExpectation::ComboTimeout(value), FlashOperationMessage::ComboTimeout(actual)) => {
-                assert_eq!(value, actual, "on flash operation at step #{step_idx}");
-            }
-            (FlashExpectation::OneShotTimeout(value), FlashOperationMessage::OneShotTimeout(actual)) => {
-                assert_eq!(value, actual, "on flash operation at step #{step_idx}");
-            }
-            (FlashExpectation::TapInterval(value), FlashOperationMessage::TapInterval(actual)) => {
-                assert_eq!(value, actual, "on flash operation at step #{step_idx}");
-            }
-            (FlashExpectation::TapCapslockInterval(value), FlashOperationMessage::TapCapslockInterval(actual)) => {
-                assert_eq!(value, actual, "on flash operation at step #{step_idx}");
-            }
-            (FlashExpectation::PriorIdleTime(value), FlashOperationMessage::PriorIdleTime(actual)) => {
-                assert_eq!(value, actual, "on flash operation at step #{step_idx}");
-            }
-            (FlashExpectation::MorseDefaultProfile(value), FlashOperationMessage::MorseDefaultProfile(actual)) => {
-                assert_eq!(value, actual, "on flash operation at step #{step_idx}");
-            }
-            (expected, actual) => {
-                panic!(
-                    "unexpected flash operation at step #{step_idx}: expected {:?}, actual {:?}",
-                    expected, actual
-                );
-            }
-        }
-    }
-
-    #[cfg(not(feature = "_no_usb"))]
+    #[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
     fn assert_report_eq(expected: Report, actual: Report, step_idx: usize) {
         use usbd_hid::descriptor::AsInputReport;
 
@@ -1384,7 +1032,7 @@ impl<'a> SimKeyboard<'a> {
         );
     }
 
-    #[cfg(not(feature = "_no_usb"))]
+    #[cfg(any(not(feature = "_no_usb"), feature = "_ble"))]
     fn assert_keyboard_state_eq(modifier: u8, mut keycodes: Vec<u8>, actual: Report, step_idx: usize) {
         let actual = match actual {
             Report::KeyboardReport(report) => report,
@@ -1413,10 +1061,6 @@ impl<'a> SimKeyboard<'a> {
 }
 
 impl SimKeyboard<'static> {
-    pub fn single_key(action: KeyAction) -> SimKeyboardBuilder<1, 1, 1, 0> {
-        Self::builder([[[action]]])
-    }
-
     pub fn builder<const ROW: usize, const COL: usize, const NUM_LAYER: usize>(
         keymap: [[[KeyAction; COL]; ROW]; NUM_LAYER],
     ) -> SimKeyboardBuilder<ROW, COL, NUM_LAYER, 0> {
@@ -1431,713 +1075,25 @@ impl SimKeyboard<'static> {
 }
 
 #[cfg(feature = "host")]
-pub struct SimHost {
-    transport: ConnectionType,
-    timeout: Duration,
-}
-
-#[cfg(feature = "host")]
-impl Default for SimHost {
-    fn default() -> Self {
-        Self {
-            transport: ConnectionType::Usb,
-            timeout: DEFAULT_TIMEOUT,
-        }
-    }
-}
+#[derive(Default)]
+pub struct SimHost;
 
 #[cfg(feature = "host")]
 impl SimHost {
-    pub fn usb() -> Self {
-        Self::default()
+    pub const fn new() -> Self {
+        Self
     }
-
-    pub fn ble() -> Self {
-        Self::default().with_transport(ConnectionType::Ble)
-    }
-
-    pub fn with_transport(mut self, transport: ConnectionType) -> Self {
-        self.transport = transport;
-        self
-    }
-
-    pub fn with_timeout(mut self, timeout: Duration) -> Self {
-        self.timeout = timeout;
-        self
-    }
-
-    #[cfg(feature = "vial")]
-    pub fn send_packet_to<'k, 'a>(&self, keyboard: &'k mut SimKeyboard<'a>, data: [u8; 32]) -> SimHostReply<'k, 'a> {
-        keyboard.enable_host();
-        SimHostReply {
-            keyboard,
-            transport: self.transport,
-            timeout: self.timeout,
-            data,
-        }
-    }
-
-    #[cfg(feature = "vial")]
-    pub fn vial<'k, 'a>(&self, keyboard: &'k mut SimKeyboard<'a>) -> SimVial<'k, 'a> {
-        keyboard.enable_host();
-        SimVial {
-            keyboard,
-            transport: self.transport,
-            timeout: self.timeout,
-        }
-    }
-
-    #[cfg(feature = "rynk")]
-    pub fn rynk<'k, 'a>(&self, keyboard: &'k mut SimKeyboard<'a>) -> SimRynk<'k, 'a> {
-        keyboard.enable_host();
-        SimRynk {
-            keyboard,
-            transport: self.transport,
-            timeout: self.timeout,
-            seq: 0,
-        }
-    }
-}
-
-#[cfg(feature = "vial")]
-#[must_use = "host transactions must end with an expectation"]
-pub struct SimHostReply<'k, 'a> {
-    keyboard: &'k mut SimKeyboard<'a>,
-    transport: ConnectionType,
-    timeout: Duration,
-    data: [u8; 32],
-}
-
-#[cfg(feature = "vial")]
-impl<'k, 'a> SimHostReply<'k, 'a> {
-    pub fn expect_ok(self) -> &'k mut SimKeyboard<'a> {
-        self.keyboard.vial_packet(
-            self.transport,
-            self.data,
-            self.timeout,
-            VialReplyExpectation::Command(self.data[0]),
-        );
-        self.keyboard
-    }
-
-    pub fn expect_echo(self) -> &'k mut SimKeyboard<'a> {
-        self.keyboard.vial_packet(
-            self.transport,
-            self.data,
-            self.timeout,
-            VialReplyExpectation::Exact(self.data),
-        );
-        self.keyboard
-    }
-
-    pub fn expect_command(self, command: u8) -> &'k mut SimKeyboard<'a> {
-        self.keyboard.vial_packet(
-            self.transport,
-            self.data,
-            self.timeout,
-            VialReplyExpectation::Command(command),
-        );
-        self.keyboard
-    }
-
-    pub fn expect(self, reply: [u8; 32]) -> &'k mut SimKeyboard<'a> {
-        self.keyboard.vial_packet(
-            self.transport,
-            self.data,
-            self.timeout,
-            VialReplyExpectation::Exact(reply),
-        );
-        self.keyboard
-    }
-}
-
-#[cfg(feature = "vial")]
-pub struct SimVial<'k, 'a> {
-    keyboard: &'k mut SimKeyboard<'a>,
-    transport: ConnectionType,
-    timeout: Duration,
-}
-
-#[cfg(feature = "vial")]
-impl<'k, 'a> SimVial<'k, 'a> {
-    pub fn raw(self, data: [u8; 32]) -> SimHostReply<'k, 'a> {
-        SimHostReply {
-            keyboard: self.keyboard,
-            transport: self.transport,
-            timeout: self.timeout,
-            data,
-        }
-    }
-
-    pub fn get_protocol_version(self) -> SimHostReply<'k, 'a> {
-        let mut data = [0u8; 32];
-        data[0] = ViaCommand::GetProtocolVersion as u8;
-        self.raw(data)
-    }
-
-    pub fn get_key(self, layer: u8, row: u8, col: u8) -> SimVialKeyReply<'k, 'a> {
-        let mut data = [0u8; 32];
-        data[0] = ViaCommand::DynamicKeymapGetKeyCode as u8;
-        data[1] = layer;
-        data[2] = row;
-        data[3] = col;
-        SimVialKeyReply { reply: self.raw(data) }
-    }
-
-    pub fn set_key(self, layer: u8, row: u8, col: u8, action: KeyAction) -> SimHostReply<'k, 'a> {
-        let mut data = [0u8; 32];
-        data[0] = ViaCommand::DynamicKeymapSetKeyCode as u8;
-        data[1] = layer;
-        data[2] = row;
-        data[3] = col;
-        data[4..6].copy_from_slice(&to_via_keycode(action).to_be_bytes());
-        self.raw(data)
-    }
-
-    pub fn get_encoder(self, layer: u8, encoder_id: u8) -> SimVialEncoderReply<'k, 'a> {
-        let mut data = [0u8; 32];
-        data[0] = ViaCommand::Vial as u8;
-        data[1] = VialCommand::GetEncoder as u8;
-        data[2] = layer;
-        data[3] = encoder_id;
-        SimVialEncoderReply { reply: self.raw(data) }
-    }
-
-    pub fn set_encoder(self, layer: u8, encoder_id: u8, action: EncoderAction) -> SimVialSetEncoderReply<'k, 'a> {
-        let mut clockwise = [0u8; 32];
-        clockwise[0] = ViaCommand::Vial as u8;
-        clockwise[1] = VialCommand::SetEncoder as u8;
-        clockwise[2] = layer;
-        clockwise[3] = encoder_id;
-        clockwise[4] = 1;
-        clockwise[5..7].copy_from_slice(&to_via_keycode(action.clockwise).to_be_bytes());
-
-        let mut counter_clockwise = [0u8; 32];
-        counter_clockwise[0] = ViaCommand::Vial as u8;
-        counter_clockwise[1] = VialCommand::SetEncoder as u8;
-        counter_clockwise[2] = layer;
-        counter_clockwise[3] = encoder_id;
-        counter_clockwise[4] = 0;
-        counter_clockwise[5..7].copy_from_slice(&to_via_keycode(action.counter_clockwise).to_be_bytes());
-
-        SimVialSetEncoderReply {
-            keyboard: self.keyboard,
-            transport: self.transport,
-            timeout: self.timeout,
-            clockwise,
-            counter_clockwise,
-        }
-    }
-
-    pub fn get_behavior_setting(self, setting: SettingKey) -> SimVialBehaviorSettingReply<'k, 'a> {
-        let mut data = [0u8; 32];
-        data[0] = ViaCommand::Vial as u8;
-        data[1] = VialCommand::GetBehaviorSetting as u8;
-        data[2..4].copy_from_slice(&(setting as u16).to_le_bytes());
-        SimVialBehaviorSettingReply { reply: self.raw(data) }
-    }
-
-    pub fn set_behavior_setting_u16(self, setting: SettingKey, value: u16) -> SimHostReply<'k, 'a> {
-        let mut data = [0u8; 32];
-        data[0] = ViaCommand::Vial as u8;
-        data[1] = VialCommand::SetBehaviorSetting as u8;
-        data[2..4].copy_from_slice(&(setting as u16).to_le_bytes());
-        data[4..6].copy_from_slice(&value.to_le_bytes());
-        self.raw(data)
-    }
-
-    pub fn set_behavior_setting_bool(self, setting: SettingKey, value: bool) -> SimHostReply<'k, 'a> {
-        let mut data = [0u8; 32];
-        data[0] = ViaCommand::Vial as u8;
-        data[1] = VialCommand::SetBehaviorSetting as u8;
-        data[2..4].copy_from_slice(&(setting as u16).to_le_bytes());
-        data[4] = value as u8;
-        self.raw(data)
-    }
-
-    pub fn get_morse(self, index: u8) -> SimVialMorseReply<'k, 'a> {
-        let mut data = [0u8; 32];
-        data[0] = ViaCommand::Vial as u8;
-        data[1] = VialCommand::DynamicEntryOp as u8;
-        data[2] = VialDynamic::DynamicVialMorseGet as u8;
-        data[3] = index;
-        SimVialMorseReply { reply: self.raw(data) }
-    }
-
-    pub fn set_morse(
-        self,
-        index: u8,
-        tap: KeyAction,
-        hold: KeyAction,
-        double_tap: KeyAction,
-        hold_after_tap: KeyAction,
-        timeout_ms: u16,
-    ) -> SimVialDynamicSetReply<'k, 'a> {
-        let mut data = [0u8; 32];
-        data[0] = ViaCommand::Vial as u8;
-        data[1] = VialCommand::DynamicEntryOp as u8;
-        data[2] = VialDynamic::DynamicVialMorseSet as u8;
-        data[3] = index;
-        data[4..6].copy_from_slice(&to_via_keycode(tap).to_le_bytes());
-        data[6..8].copy_from_slice(&to_via_keycode(hold).to_le_bytes());
-        data[8..10].copy_from_slice(&to_via_keycode(double_tap).to_le_bytes());
-        data[10..12].copy_from_slice(&to_via_keycode(hold_after_tap).to_le_bytes());
-        data[12..14].copy_from_slice(&timeout_ms.to_le_bytes());
-        SimVialDynamicSetReply { reply: self.raw(data) }
-    }
-
-    pub fn get_combo(self, index: u8) -> SimVialComboReply<'k, 'a> {
-        let mut data = [0u8; 32];
-        data[0] = ViaCommand::Vial as u8;
-        data[1] = VialCommand::DynamicEntryOp as u8;
-        data[2] = VialDynamic::DynamicVialComboGet as u8;
-        data[3] = index;
-        SimVialComboReply { reply: self.raw(data) }
-    }
-
-    pub fn set_combo<const N: usize>(
-        self,
-        index: u8,
-        actions: [KeyAction; N],
-        output: KeyAction,
-    ) -> SimVialDynamicSetReply<'k, 'a> {
-        assert!(
-            N <= crate::COMBO_MAX_LENGTH,
-            "simulator combo helper received too many actions"
-        );
-
-        let mut data = [0u8; 32];
-        data[0] = ViaCommand::Vial as u8;
-        data[1] = VialCommand::DynamicEntryOp as u8;
-        data[2] = VialDynamic::DynamicVialComboSet as u8;
-        data[3] = index;
-        for (idx, action) in actions.into_iter().enumerate() {
-            let start = 4 + idx * 2;
-            data[start..start + 2].copy_from_slice(&to_via_keycode(action).to_le_bytes());
-        }
-        let output_start = 4 + crate::COMBO_MAX_LENGTH * 2;
-        data[output_start..output_start + 2].copy_from_slice(&to_via_keycode(output).to_le_bytes());
-
-        SimVialDynamicSetReply { reply: self.raw(data) }
-    }
-
-    pub fn unsupported_dynamic_entry(self) -> SimHostReply<'k, 'a> {
-        let mut data = [0u8; 32];
-        data[0] = ViaCommand::Vial as u8;
-        data[1] = VialCommand::DynamicEntryOp as u8;
-        data[2] = VialDynamic::Unhandled as u8;
-        self.raw(data)
-    }
-}
-
-#[cfg(feature = "vial")]
-#[must_use = "Vial requests must end with an expectation"]
-pub struct SimVialKeyReply<'k, 'a> {
-    reply: SimHostReply<'k, 'a>,
-}
-
-#[cfg(feature = "vial")]
-impl<'k, 'a> SimVialKeyReply<'k, 'a> {
-    pub fn expect(self, action: KeyAction) -> &'k mut SimKeyboard<'a> {
-        let mut expected = self.reply.data;
-        expected[4..6].copy_from_slice(&to_via_keycode(action).to_be_bytes());
-        self.reply.expect(expected)
-    }
-}
-
-#[cfg(feature = "vial")]
-#[must_use = "Vial requests must end with an expectation"]
-pub struct SimVialEncoderReply<'k, 'a> {
-    reply: SimHostReply<'k, 'a>,
-}
-
-#[cfg(feature = "vial")]
-impl<'k, 'a> SimVialEncoderReply<'k, 'a> {
-    pub fn expect(self, action: EncoderAction) -> &'k mut SimKeyboard<'a> {
-        let mut expected = [0u8; 32];
-        expected[0..2].copy_from_slice(&to_via_keycode(action.counter_clockwise).to_be_bytes());
-        expected[2..4].copy_from_slice(&to_via_keycode(action.clockwise).to_be_bytes());
-        self.reply.expect(expected)
-    }
-}
-
-#[cfg(feature = "vial")]
-#[must_use = "Vial requests must end with an expectation"]
-pub struct SimVialSetEncoderReply<'k, 'a> {
-    keyboard: &'k mut SimKeyboard<'a>,
-    transport: ConnectionType,
-    timeout: Duration,
-    clockwise: [u8; 32],
-    counter_clockwise: [u8; 32],
-}
-
-#[cfg(feature = "vial")]
-impl<'k, 'a> SimVialSetEncoderReply<'k, 'a> {
-    pub fn expect_ok(self) -> &'k mut SimKeyboard<'a> {
-        self.keyboard.vial_packet(
-            self.transport,
-            self.clockwise,
-            self.timeout,
-            VialReplyExpectation::Exact(self.clockwise),
-        );
-        self.keyboard.vial_packet(
-            self.transport,
-            self.counter_clockwise,
-            self.timeout,
-            VialReplyExpectation::Exact(self.counter_clockwise),
-        );
-        self.keyboard
-    }
-}
-
-#[cfg(feature = "vial")]
-#[must_use = "Vial requests must end with an expectation"]
-pub struct SimVialBehaviorSettingReply<'k, 'a> {
-    reply: SimHostReply<'k, 'a>,
-}
-
-#[cfg(feature = "vial")]
-impl<'k, 'a> SimVialBehaviorSettingReply<'k, 'a> {
-    pub fn expect_u16(self, value: u16) -> &'k mut SimKeyboard<'a> {
-        let mut expected = [0xFF; 32];
-        expected[0] = 0;
-        expected[1..3].copy_from_slice(&value.to_le_bytes());
-        self.reply.expect(expected)
-    }
-
-    pub fn expect_bool(self, value: bool) -> &'k mut SimKeyboard<'a> {
-        let mut expected = [0xFF; 32];
-        expected[0] = 0;
-        expected[1] = value as u8;
-        self.reply.expect(expected)
-    }
-}
-
-#[cfg(feature = "vial")]
-#[must_use = "Vial requests must end with an expectation"]
-pub struct SimVialMorseReply<'k, 'a> {
-    reply: SimHostReply<'k, 'a>,
-}
-
-#[cfg(feature = "vial")]
-impl<'k, 'a> SimVialMorseReply<'k, 'a> {
-    pub fn expect(
-        self,
-        tap: KeyAction,
-        hold: KeyAction,
-        double_tap: KeyAction,
-        hold_after_tap: KeyAction,
-        timeout_ms: u16,
-    ) -> &'k mut SimKeyboard<'a> {
-        let mut expected = self.reply.data;
-        expected[0] = 0;
-        expected[1..3].copy_from_slice(&to_via_keycode(tap).to_le_bytes());
-        expected[3..5].copy_from_slice(&to_via_keycode(hold).to_le_bytes());
-        expected[5..7].copy_from_slice(&to_via_keycode(double_tap).to_le_bytes());
-        expected[7..9].copy_from_slice(&to_via_keycode(hold_after_tap).to_le_bytes());
-        expected[9..11].copy_from_slice(&timeout_ms.to_le_bytes());
-        self.reply.expect(expected)
-    }
-}
-
-#[cfg(feature = "vial")]
-#[must_use = "Vial requests must end with an expectation"]
-pub struct SimVialComboReply<'k, 'a> {
-    reply: SimHostReply<'k, 'a>,
-}
-
-#[cfg(feature = "vial")]
-impl<'k, 'a> SimVialComboReply<'k, 'a> {
-    pub fn expect<const N: usize>(self, actions: [KeyAction; N], output: KeyAction) -> &'k mut SimKeyboard<'a> {
-        assert!(
-            N <= crate::COMBO_MAX_LENGTH,
-            "simulator combo helper received too many actions"
-        );
-
-        let mut expected = self.reply.data;
-        expected[0] = 0;
-        for (idx, action) in actions.into_iter().enumerate() {
-            let start = 1 + idx * 2;
-            expected[start..start + 2].copy_from_slice(&to_via_keycode(action).to_le_bytes());
-        }
-        let output_start = 1 + crate::COMBO_MAX_LENGTH * 2;
-        expected[output_start..output_start + 2].copy_from_slice(&to_via_keycode(output).to_le_bytes());
-        self.reply.expect(expected)
-    }
-}
-
-#[cfg(feature = "vial")]
-#[must_use = "Vial requests must end with an expectation"]
-pub struct SimVialDynamicSetReply<'k, 'a> {
-    reply: SimHostReply<'k, 'a>,
-}
-
-#[cfg(feature = "vial")]
-impl<'k, 'a> SimVialDynamicSetReply<'k, 'a> {
-    pub fn expect_ok(self) -> &'k mut SimKeyboard<'a> {
-        let mut expected = self.reply.data;
-        expected[0] = 0;
-        self.reply.expect(expected)
-    }
-}
-
-#[cfg(feature = "rynk")]
-pub struct SimRynk<'k, 'a> {
-    keyboard: &'k mut SimKeyboard<'a>,
-    transport: ConnectionType,
-    timeout: Duration,
-    seq: u8,
-}
-
-#[cfg(feature = "rynk")]
-impl<'k, 'a> SimRynk<'k, 'a> {
-    pub fn with_seq(mut self, seq: u8) -> Self {
-        self.seq = seq;
-        self
-    }
-
-    pub fn request<E: Endpoint>(self, payload: E::Request) -> SimRynkReply<'k, 'a, E> {
-        SimRynkReply {
-            keyboard: self.keyboard,
-            transport: self.transport,
-            timeout: self.timeout,
-            seq: self.seq,
-            request: rynk_request_frame(E::CMD, self.seq, &payload),
-            endpoint: PhantomData,
-        }
-    }
-
-    pub fn get_version(self) -> SimRynkReply<'k, 'a, command::GetVersion> {
-        self.request::<command::GetVersion>(())
-    }
-
-    pub fn get_key(self, layer: u8, row: u8, col: u8) -> SimRynkReply<'k, 'a, command::GetKeyAction> {
-        self.request::<command::GetKeyAction>(rmk_types::protocol::rynk::KeyPosition { layer, row, col })
-    }
-
-    pub fn set_key(
-        self,
-        layer: u8,
-        row: u8,
-        col: u8,
-        action: KeyAction,
-    ) -> SimRynkReply<'k, 'a, command::SetKeyAction> {
-        self.request::<command::SetKeyAction>(rmk_types::protocol::rynk::SetKeyRequest {
-            position: rmk_types::protocol::rynk::KeyPosition { layer, row, col },
-            action,
-        })
-    }
-
-    pub fn get_encoder(self, layer: u8, encoder_id: u8) -> SimRynkReply<'k, 'a, command::GetEncoderAction> {
-        self.request::<command::GetEncoderAction>(rmk_types::protocol::rynk::GetEncoderRequest { encoder_id, layer })
-    }
-
-    pub fn set_encoder(
-        self,
-        layer: u8,
-        encoder_id: u8,
-        action: EncoderAction,
-    ) -> SimRynkReply<'k, 'a, command::SetEncoderAction> {
-        self.request::<command::SetEncoderAction>(rmk_types::protocol::rynk::SetEncoderRequest {
-            encoder_id,
-            layer,
-            action,
-        })
-    }
-}
-
-#[cfg(feature = "rynk")]
-#[must_use = "Rynk requests must end with an expectation"]
-pub struct SimRynkReply<'k, 'a, E: Endpoint> {
-    keyboard: &'k mut SimKeyboard<'a>,
-    transport: ConnectionType,
-    timeout: Duration,
-    seq: u8,
-    request: Vec<u8>,
-    endpoint: PhantomData<E>,
-}
-
-#[cfg(feature = "rynk")]
-impl<'k, 'a, E: Endpoint> SimRynkReply<'k, 'a, E> {
-    pub fn expect(self, response: E::Response) -> &'k mut SimKeyboard<'a> {
-        let expected = rynk_response_frame(E::CMD, self.seq, &response);
-        self.keyboard
-            .rynk_packet(self.transport, self.timeout, self.request, expected);
-        self.keyboard
-    }
-
-    pub fn expect_ok(self) -> &'k mut SimKeyboard<'a>
-    where
-        E: Endpoint<Response = ()>,
-    {
-        self.expect(())
-    }
-
-    pub fn expect_error(self, error: RynkError) -> &'k mut SimKeyboard<'a> {
-        let expected = rynk_error_response_frame(E::CMD, self.seq, error);
-        self.keyboard
-            .rynk_packet(self.transport, self.timeout, self.request, expected);
-        self.keyboard
-    }
-}
-
-#[cfg(feature = "rynk")]
-fn rynk_request_frame<T: serde::Serialize>(cmd: Cmd, seq: u8, payload: &T) -> Vec<u8> {
-    let mut buf = std::vec![0u8; rmk_types::constants::RYNK_BUFFER_SIZE];
-    RynkMessage::build(&mut buf, cmd, seq, payload).expect("simulator Rynk request should encode");
-    buf
-}
-
-#[cfg(feature = "rynk")]
-fn rynk_response_frame<T: serde::Serialize>(cmd: Cmd, seq: u8, payload: &T) -> Vec<u8> {
-    let mut buf = std::vec![0u8; rmk_types::constants::RYNK_BUFFER_SIZE];
-    let msg = RynkMessage::build(&mut buf, cmd, seq, &Ok::<&T, RynkError>(payload))
-        .expect("simulator Rynk response should encode");
-    let frame_len = msg.frame_len();
-    buf.truncate(frame_len);
-    buf
-}
-
-#[cfg(feature = "rynk")]
-fn rynk_error_response_frame(cmd: Cmd, seq: u8, error: RynkError) -> Vec<u8> {
-    let mut buf = std::vec![0u8; rmk_types::constants::RYNK_BUFFER_SIZE];
-    let msg = RynkMessage::build(&mut buf, cmd, seq, &Err::<(), RynkError>(error))
-        .expect("simulator Rynk error response should encode");
-    let frame_len = msg.frame_len();
-    buf.truncate(frame_len);
-    buf
 }
 
 #[cfg(feature = "storage")]
-pub mod flash {
-    use std::sync::{Arc, Mutex};
-    use std::vec::Vec;
+pub mod flash;
+#[cfg(feature = "rynk")]
+pub mod rynk;
+#[cfg(feature = "vial")]
+pub mod vial;
 
-    use embedded_storage::nor_flash::{
-        ErrorType, NorFlash, NorFlashError, NorFlashErrorKind, ReadNorFlash, check_erase, check_read, check_write,
-    };
+mod executor;
 
-    #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-    pub enum InMemoryFlashError {
-        OutOfBounds,
-        NotAligned,
-        WriteRequiresErase,
-        Poisoned,
-    }
-
-    impl NorFlashError for InMemoryFlashError {
-        fn kind(&self) -> NorFlashErrorKind {
-            match self {
-                Self::OutOfBounds => NorFlashErrorKind::OutOfBounds,
-                Self::NotAligned => NorFlashErrorKind::NotAligned,
-                Self::WriteRequiresErase | Self::Poisoned => NorFlashErrorKind::Other,
-            }
-        }
-    }
-
-    #[derive(Debug, Clone)]
-    pub struct InMemoryFlash<const SIZE: usize, const ERASE: usize = 4096, const WRITE: usize = 4> {
-        data: Arc<Mutex<Vec<u8>>>,
-    }
-
-    impl<const SIZE: usize, const ERASE: usize, const WRITE: usize> Default for InMemoryFlash<SIZE, ERASE, WRITE> {
-        fn default() -> Self {
-            Self::new()
-        }
-    }
-
-    impl<const SIZE: usize, const ERASE: usize, const WRITE: usize> InMemoryFlash<SIZE, ERASE, WRITE> {
-        pub fn new() -> Self {
-            Self {
-                data: Arc::new(Mutex::new(std::vec![0xFF; SIZE])),
-            }
-        }
-
-        pub fn snapshot(&self) -> Vec<u8> {
-            self.data.lock().expect("in-memory flash mutex poisoned").clone()
-        }
-
-        fn map_error(kind: NorFlashErrorKind) -> InMemoryFlashError {
-            match kind {
-                NorFlashErrorKind::OutOfBounds => InMemoryFlashError::OutOfBounds,
-                NorFlashErrorKind::NotAligned => InMemoryFlashError::NotAligned,
-                _ => InMemoryFlashError::Poisoned,
-            }
-        }
-    }
-
-    impl<const SIZE: usize, const ERASE: usize, const WRITE: usize> ErrorType for InMemoryFlash<SIZE, ERASE, WRITE> {
-        type Error = InMemoryFlashError;
-    }
-
-    impl<const SIZE: usize, const ERASE: usize, const WRITE: usize> ReadNorFlash for InMemoryFlash<SIZE, ERASE, WRITE> {
-        const READ_SIZE: usize = 1;
-
-        fn read(&mut self, offset: u32, bytes: &mut [u8]) -> Result<(), Self::Error> {
-            check_read(self, offset, bytes.len()).map_err(Self::map_error)?;
-            let data = self.data.lock().map_err(|_| InMemoryFlashError::Poisoned)?;
-            let offset = offset as usize;
-            bytes.copy_from_slice(&data[offset..offset + bytes.len()]);
-            Ok(())
-        }
-
-        fn capacity(&self) -> usize {
-            SIZE
-        }
-    }
-
-    impl<const SIZE: usize, const ERASE: usize, const WRITE: usize> NorFlash for InMemoryFlash<SIZE, ERASE, WRITE> {
-        const WRITE_SIZE: usize = WRITE;
-        const ERASE_SIZE: usize = ERASE;
-
-        fn erase(&mut self, from: u32, to: u32) -> Result<(), Self::Error> {
-            check_erase(self, from, to).map_err(Self::map_error)?;
-            let mut data = self.data.lock().map_err(|_| InMemoryFlashError::Poisoned)?;
-            data[from as usize..to as usize].fill(0xFF);
-            Ok(())
-        }
-
-        fn write(&mut self, offset: u32, bytes: &[u8]) -> Result<(), Self::Error> {
-            check_write(self, offset, bytes.len()).map_err(Self::map_error)?;
-            let mut data = self.data.lock().map_err(|_| InMemoryFlashError::Poisoned)?;
-            let offset = offset as usize;
-            for (idx, byte) in bytes.iter().enumerate() {
-                let current = data[offset + idx];
-                if current & byte != *byte {
-                    return Err(InMemoryFlashError::WriteRequiresErase);
-                }
-                data[offset + idx] = current & byte;
-            }
-            Ok(())
-        }
-    }
-
-    impl<const SIZE: usize, const ERASE: usize, const WRITE: usize> embedded_storage_async::nor_flash::ReadNorFlash
-        for InMemoryFlash<SIZE, ERASE, WRITE>
-    {
-        const READ_SIZE: usize = 1;
-
-        async fn read(&mut self, offset: u32, bytes: &mut [u8]) -> Result<(), Self::Error> {
-            ReadNorFlash::read(self, offset, bytes)
-        }
-
-        fn capacity(&self) -> usize {
-            SIZE
-        }
-    }
-
-    impl<const SIZE: usize, const ERASE: usize, const WRITE: usize> embedded_storage_async::nor_flash::NorFlash
-        for InMemoryFlash<SIZE, ERASE, WRITE>
-    {
-        const WRITE_SIZE: usize = WRITE;
-        const ERASE_SIZE: usize = ERASE;
-
-        async fn erase(&mut self, from: u32, to: u32) -> Result<(), Self::Error> {
-            NorFlash::erase(self, from, to)
-        }
-
-        async fn write(&mut self, offset: u32, bytes: &[u8]) -> Result<(), Self::Error> {
-            NorFlash::write(self, offset, bytes)
-        }
-    }
+pub fn test_block_on<F: core::future::Future>(future: F) -> F::Output {
+    executor::test_block_on(future)
 }
