+++ rmk/tests/keyboard_morse_permissive_hold_test.rs
@@ -9,11 +9,11 @@ use crate::common::{KC_LGUI, KC_LSHIFT, TEST_KEYMAP};
 
 const PERMISSIVE_HOLD_PROFILE: MorseProfile =
     MorseProfile::new(Some(false), Some(MorseMode::PermissiveHold), Some(250u16), Some(250u16));
-const PERMISSIVE_HOLD_SETUP: SimKeyboardSetup<5, 14> = SIMPLE_MORSE_SETUP.morse_profile(PERMISSIVE_HOLD_PROFILE);
+const PERMISSIVE_HOLD_SETUP: SimKeyboardSetup = SIMPLE_MORSE_SETUP.morse_profile(PERMISSIVE_HOLD_PROFILE);
 
 #[test]
 fn test_tap() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -33,7 +33,7 @@ fn test_tap() {
 
 #[test]
 fn test_hold() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -53,7 +53,7 @@ fn test_hold() {
 
 #[test]
 fn test_mt_1() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -79,7 +79,7 @@ fn test_mt_1() {
 
 #[test]
 fn test_mt_2() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -105,7 +105,7 @@ fn test_mt_2() {
 
 #[test]
 fn test_mt_3() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -131,7 +131,7 @@ fn test_mt_3() {
 
 #[test]
 fn test_mt_4() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -157,7 +157,7 @@ fn test_mt_4() {
 
 #[test]
 fn test_mt_5() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -183,7 +183,7 @@ fn test_mt_5() {
 
 #[test]
 fn test_mt_6() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -209,7 +209,7 @@ fn test_mt_6() {
 
 #[test]
 fn test_mt_timeout_1() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -235,7 +235,7 @@ fn test_mt_timeout_1() {
 
 #[test]
 fn test_mt_timeout_2() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -261,7 +261,7 @@ fn test_mt_timeout_2() {
 
 #[test]
 fn test_mt_timeout_3() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -287,7 +287,7 @@ fn test_mt_timeout_3() {
 
 #[test]
 fn test_mt_timeout_4() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -313,7 +313,7 @@ fn test_mt_timeout_4() {
 
 #[test]
 fn test_mt_timeout_5() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -339,7 +339,7 @@ fn test_mt_timeout_5() {
 
 #[test]
 fn test_mt_timeout_6() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -365,7 +365,7 @@ fn test_mt_timeout_6() {
 
 #[test]
 fn test_mt_timeout_7() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -391,7 +391,7 @@ fn test_mt_timeout_7() {
 
 #[test]
 fn test_mt_timeout_8() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -417,7 +417,7 @@ fn test_mt_timeout_8() {
 
 #[test]
 fn test_mt_timeout_9() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -443,7 +443,7 @@ fn test_mt_timeout_9() {
 
 #[test]
 fn test_mt_timeout_10() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -469,7 +469,7 @@ fn test_mt_timeout_10() {
 
 #[test]
 fn test_morse_lt_1() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -493,7 +493,7 @@ fn test_morse_lt_1() {
 
 #[test]
 fn test_morse_lt_2() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -519,7 +519,7 @@ fn test_morse_lt_2() {
 
 #[test]
 fn test_morse_lt_3() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -545,7 +545,7 @@ fn test_morse_lt_3() {
 
 #[test]
 fn test_morse_lt_4() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -571,7 +571,7 @@ fn test_morse_lt_4() {
 
 #[test]
 fn test_morse_lt_5() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -597,7 +597,7 @@ fn test_morse_lt_5() {
 
 #[test]
 fn test_morse_lt_6() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -623,7 +623,7 @@ fn test_morse_lt_6() {
 
 #[test]
 fn test_morse_lt_timeout_1() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -647,7 +647,7 @@ fn test_morse_lt_timeout_1() {
 
 #[test]
 fn test_morse_lt_timeout_2() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -671,7 +671,7 @@ fn test_morse_lt_timeout_2() {
 
 #[test]
 fn test_morse_lt_timeout_3() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -695,7 +695,7 @@ fn test_morse_lt_timeout_3() {
 
 #[test]
 fn test_morse_lt_timeout_4() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -719,7 +719,7 @@ fn test_morse_lt_timeout_4() {
 
 #[test]
 fn test_morse_lt_timeout_5() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -743,7 +743,7 @@ fn test_morse_lt_timeout_5() {
 
 #[test]
 fn test_morse_lt_timeout_6() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -767,7 +767,7 @@ fn test_morse_lt_timeout_6() {
 
 #[test]
 fn test_morse_lt_timeout_7() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -791,7 +791,7 @@ fn test_morse_lt_timeout_7() {
 
 #[test]
 fn test_morse_lt_timeout_8() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -815,7 +815,7 @@ fn test_morse_lt_timeout_8() {
 
 #[test]
 fn test_morse_lt_timeout_9() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -839,7 +839,7 @@ fn test_morse_lt_timeout_9() {
 
 #[test]
 fn test_morse_lt_timeout_10() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -863,7 +863,7 @@ fn test_morse_lt_timeout_10() {
 
 #[test]
 fn test_trigger() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -889,7 +889,7 @@ fn test_trigger() {
 
 #[test]
 fn test_with_combo_1() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .combos_global(MORSE_2_KEY_COMBOS)
@@ -918,7 +918,7 @@ fn test_with_combo_1() {
 
 #[test]
 fn test_with_combo_2() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .combos_global(MORSE_2_KEY_COMBOS)
@@ -945,7 +945,7 @@ fn test_with_combo_2() {
 
 #[test]
 fn test_with_combo_3() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .combos_global(MORSE_2_KEY_COMBOS)
@@ -972,7 +972,7 @@ fn test_with_combo_3() {
 
 #[test]
 fn test_with_combo_4() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .combos_global(MORSE_2_KEY_COMBOS)
@@ -1001,7 +1001,7 @@ fn test_with_combo_4() {
 
 #[test]
 fn test_with_combo_5() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .combos_global(MORSE_2_KEY_COMBOS)
@@ -1028,7 +1028,7 @@ fn test_with_combo_5() {
 
 #[test]
 fn test_with_combo_6() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .combos_global(MORSE_2_KEY_COMBOS)
@@ -1063,7 +1063,7 @@ fn test_with_combo_6() {
 
 #[test]
 fn test_with_combo_7() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .combos_global(MORSE_2_KEY_COMBOS)
@@ -1094,7 +1094,7 @@ fn test_with_combo_7() {
 
 #[test]
 fn test_with_combo_8() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .combos_global(MORSE_2_KEY_COMBOS)
@@ -1127,7 +1127,7 @@ fn test_with_combo_8() {
 
 #[test]
 fn test_timeout() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -1153,7 +1153,7 @@ fn test_timeout() {
 
 #[test]
 fn test_quick_tap() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -1179,7 +1179,7 @@ fn test_quick_tap() {
 
 #[test]
 fn test_multi_tap() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -1211,7 +1211,7 @@ fn test_multi_tap() {
 
 #[test]
 fn test_layer_tap() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -1251,7 +1251,7 @@ fn test_layer_tap() {
 
 #[test]
 fn test_rolling_with_layer_tap() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -1299,7 +1299,7 @@ fn test_rolling_with_layer_tap() {
 
 #[test]
 fn test_timeout_rolled_release() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -1325,7 +1325,7 @@ fn test_timeout_rolled_release() {
 
 #[test]
 fn test_timeout_rolled_release_2() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -1351,7 +1351,7 @@ fn test_timeout_rolled_release_2() {
 
 #[test]
 fn test_timeout_and_release() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -1377,7 +1377,7 @@ fn test_timeout_and_release() {
 
 #[test]
 fn test_timeout_and_release_with_other_morse_key() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -1403,7 +1403,7 @@ fn test_timeout_and_release_with_other_morse_key() {
 
 #[test]
 fn test_rolling_release_order() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -1435,7 +1435,7 @@ fn test_rolling_release_order() {
 
 #[test]
 fn test_rolling_release_order_2() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -1467,7 +1467,7 @@ fn test_rolling_release_order_2() {
 
 #[test]
 fn test_rolling_release_order_3() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -1499,7 +1499,7 @@ fn test_rolling_release_order_3() {
 
 #[test]
 fn test_multiple_permissive_hold() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -1531,7 +1531,7 @@ fn test_multiple_permissive_hold() {
 
 #[test]
 fn test_complex_rolling() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -1569,7 +1569,7 @@ fn test_complex_rolling() {
 
 #[test]
 fn test_flow_tap() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
@@ -1602,7 +1602,7 @@ fn test_flow_tap() {
 // Ref: https://github.com/HaoboGu/rmk/pull/496
 #[test]
 fn test_previous_rolling_keypress() {
-    crate::common::test_block_on::test_block_on(async {
+    crate::common::test_block_on(async {
         let mut keyboard = SimKeyboard::builder(TEST_KEYMAP)
             .setup(PERMISSIVE_HOLD_SETUP)
             .build()
